# www/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Barbearia API",
        "version": "0.1"
    },
    "paths": {
        "/api/barbearia": {
            "get": {
                "tags": [
                    "Barbearia"
                ],
                "summary": "Get list of barbearia",
                "description": "Returns list of barbearia",
                "operationId": "getbarbeariaList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Barbearia"
                ],
                "summary": "Add barbearia",
                "operationId": "addbarbearia",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "localizacao": {
                                        "type": "string"
                                    },
                                    "avatar": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/barbearia/{id}": {
            "put": {
                "tags": [
                    "Barbearia"
                ],
                "summary": "Update existing Barbearia",
                "description": "Returns updated Barbearia data",
                "operationId": "putBarbearia",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "localizacao": {
                                        "type": "string"
                                    },
                                    "avatar": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid ID"
                    },
                    "401": {
                        "description": "Invalid data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Barbearia"
                ],
                "summary": "Delete barbearia",
                "operationId": "deleteBarbearia",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Barbearia id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/planos": {
            "get": {
                "tags": [
                    "Planos"
                ],
                "summary": "Get list of Planos",
                "description": "Returns list of Planos",
                "operationId": "getPlanosList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Planos"
                ],
                "summary": "Add Planos",
                "operationId": "addPlanos",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "periodo": {
                                        "type": "integer"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "valor": {
                                        "type": "string",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "202": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/planos/{id}": {
            "put": {
                "tags": [
                    "Planos"
                ],
                "summary": "Update existing Planos",
                "description": "Returns updated Planos data",
                "operationId": "putPlanos",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "periodo": {
                                        "type": "integer"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "valor": {
                                        "type": "string",
                                        "format": "decimal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid ID"
                    },
                    "202": {
                        "description": "Invalid data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Planos"
                ],
                "summary": "Delete Planos",
                "operationId": "deletePlanos",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Planos id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}